requirements.txt
beautifulsoup4==4.9.3
requests==2.25.1
selenium
python-dotenv
sentry-sdk==1.1.0
pytest-mock==3.6.1
pytest-sugar
lxml==4.6.3
pydantic==1.8.2
pytest==6.2.4
sentry-sdk==1.1.0
raven==6.10.0
faker==8.10.3
boto3==1.15.0
coralogix_logger==2.0.4
